+++ packages/resin-universal-mixpanel/src/browser.js
@@ -4,8 +4,8 @@ var TRACKER_NAME = 'resinAnalytics'
 
 // normalize the API to match the one of the node module
 module.exports = {
-	init: function(token) {
-		mixpanel.init(token, {}, TRACKER_NAME)
+	init: function(token, options) {
+		mixpanel.init(token, options, TRACKER_NAME)
 		return mixpanel[TRACKER_NAME]
 	}
 }