+++ packages/resin-universal-mixpanel/CHANGELOG.md
@@ -1,3 +1,5 @@
+* Accept mixpanel options passed to .init()
+
 # 2.0.0
 
 * **Breaking!** Mixpanel is now returned synchronously in Node and the Browser, not wrapped in a promise.